pyproject.toml
[project]
name = "ride-service"
version = "0.1.0"
description = "Service with ride endpoints"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "firebase-admin>=6.6.0",
    "google-maps-routing>=0.6.15",
    "polyline>=2.0.2",
    "pydantic-settings>=2.7.1",
    "uvicorn>=0.34.0",
]
